Seamless Local EC2 Database Backup to S3 Using IAM Role Integration!
This documentation explains how to create a backup of an locally cretated insce database and store it securely in an S3 bucket. The process involves generating a database dump, configuring AWS CLI, and uploading the backup to S3.

Key highlights include:
Using mysqldump to back up the database.
Storing the backup in S3 for durability and scalability.